MXS-4616 Set 1000 as the maximum number of statements to be executed
This option is adjustable in the query setting dialog.

This uses a ready-made MIT library called `dbgate-query-splitter`
for splitting query. This library supports MySQL dialect and
custom delimiter. For now it's used only for counting the
number of statements. It won't return an accurate number for
query contains compound statements unless it includes the changes
of `DELIMITER`. However, the default max statements is 1000, so even
it's not perfect, it should prevent the GUI from running of memory.

In the future, we may require users to include the command that
modifies the `DELIMITER` in the query. This would ensure that the
splitting function can accurately divide the query.