MXS-4734: Fix SET TRANSACTION routing
The query must not be classified as a session command as it only
modifies the behavior of the next transaction. This also means that it
cannot be treated as a read-only transaction and the safe thing to do is
to route it to the current master.

Modified the tests to expect this behavior and added a new test case
that reproduces the problem and verifies that it is fixed.