Report completely unknown commands
If some input that does not match the grammar at all is given,
x3::phrase_parse returns false but it does not produce an error or an
exception. In these situations a custom message is needed to make sure
something is always returned to the client.

This caused a debug assertion in 23.08 where the error generation
functions expect the error message to not be empty.