MXS-4735: Add connection IDs to handleError messages
For some database drivers or ORM frameworks, it is either difficult or
impossible to extract the thread ID of the connection where the error
occurred. Since the session ID in MaxScale and the connection ID to the
backend server are necessary for post mortem analysis of connection
failures, having this information be readily available in the error
messages will simplify that task.