MXS-4725: Cache resolved symbols
This significantly speeds up the process of generating a
stacktrace. With 1000 repeated calls of the stacktrace generation
endpoint and without caching, about 85% of the CPU time is spent
resolving the symbol names. With the caching enabled, the symbol
resolution does not even show up in the profiling.

The cache size is set to a fixed size of 8192 elements. If the size is
exceeded, the whole cache is cleared. Preliminary testing shows that
this is adequately large to cover most sampling needs and the cache ends
up not being cleared when used with a basic set of modules.