MXS-4725: Add `maxctrl debug stacktrace`
The default behavior of the command is to take a single stacktrace
snapshot and group the results with the most common stacktraces
first. This can be used to observe what a thread in MaxScale is doing.

With the --duration option, the command can be used to aggregate samples
over a period of time to get a human-readable profile of the most common
stacktraces. This format can be used for a simple performance profile
that does not need post-processing.

If the --duration option is combined with the --fold option, the
stacktraces are folded and formatted in a way that is directly
compatible with the FlameGraph utility. This way the output of MaxCtrl
can fed either directly into the script that processes the folded
stacktraces into a SVG file or into a text file for later off-site
processing.