MXS-4264: Optimize luafilter API usage
The API now uses references to the routeQuery and clientReply
functions. This removes the global lookup which is a relatively large
cost when it is known that the values should never change. Not passing
any arguments to the functions also delays the cost to the point at
which the values are needed. This way the cost of managing the data is
only paid when needed.

Updated the documentation to reflect the changes in the API and added a
limitations section.