MXS-4709 Add contact and documentation items in sidebar
According to the UX guideline, permanent link for document and
contact items are placed at the bottom of the sidebar. Although
this task is supposed to add only the contact item but this also
adds a document item to make the UI looks balance.

Clicking on the document item will open a new browser tab
to maxscale documentation. e.g.  `https://mariadb.com/kb/en/mariadb-maxscale-23-02/`