Take mxb::small_vector into use
Readwritesplit used a std::vector to sort a handful of pointers. By
swapping it out for a mxb::small_vector, the number of heap allocations
drops down to zero if the number of backend servers remains below 6. The
choice of how many values to store internally was mostly arbitrary but
with 6 values the size of the Candidates type ends up being exactly 64
bytes.

The mxs::Backend class stores the types of responses it tracks but in
normal usage there's going to be only one response to track. The
std::deque was a somewhat bad choice for this as it causes two
dereferences which start to take a comparatively long time. A
std::vector would be an adequate replacement but with a
mxb::small_vector, the memory allocations are avoided completely.