MXS-3986 Fix startup schroedinbug
At initial startup where no binlog data has yet
been saved, a Reader can startup before the Writer has
written any binlog data. In that case the Reader will
check if the empty gtid it has is included in the empty
gtid the Writer has, which obviously is true in set theory,
so it merrily goes ahead and tries to open an empty file.
The chance of this happening has apparently gone from ~0%
to ~100% today.