Collect all parsing info into QueryInfo
The overhead of having to go throught the Parser and Helper classes to
get about one byte's worth of data is starting to become a costly
overhead. In addition, the overhead of the QCInfoCache instantiation for
each function, even if already cached, is starting to show up as
something that takes CPU time in QueryClassifier::update_route_info().

By collecting all the information in one function call into a common
struct, the various overheads are minimized without having to implicitly
which Parser or Helper is being used. This seems to reduce total time
spent in update_route_info() by about half which also showed up as a
2.5% increase in QPS in one test run.