MXS-3986 Handle gtid search of the very first file
This change is a step towards searching multiple gtids
(or domains) at the same time. This is rather academic
but helps testing.

If a binlog has an empty gtid list, i.e. it is the
very first binlog, generate a gtid list for that
case. The file is currently always short.