Fix Wshadow=local errors on Debian 10
The redundant type aliases in the authenticators caused an error that
newer GCC versions do not warn about.

There was also a static class variable in the mock filter framework that
was seen as a local.

Somewhere in the Spirit X3 internals there's a template that ends up
declaring a type that end up shadowing some part of another template.