MXS-4755 Generate list of fixed internal and external CVEs
For Jira bugs and tasks it will be checked whether they have
a label that starts with 'CVE-', in which case the label is
assumed to be a CVE id.

In case of a bug, the bug fix is assumed to resolve a CVE
that is related in MaxScale itself.

In case of a task, the closed task is assumed to resolve a
CVE that is not related to MaxScale itself, but e.g. to a
library that MaxScale is using. The CVE may be resolved, for
instance, by a library upgrade.