MXS-4756 Fix dashboard servers caching issue
The tableRows computed property fails to update the rows in the
table if getAllMonitorsMap returns an empty object. The function
stops when accessing an undefined property of the monitor, as a result,
it doesn't show the new data accurately. This case occurs in an edge
case where the users are logged out by the application after receiving
401, the data for servers and monitors are cached.

This adds a fallback object for accessing monitor attributes