MXS-4762: Make type mismatch errors more exact
The errors for mismatching or unknown object types should only be sent
to the client if the endpoint would otherwise match. This means that a
`GET /services/asdf/relationships/filters` will return an error about
`asdf` not being a service but `GET /services/asdf/asdf/asdf` will
return a normal 404 Not Found error.

The previous code wrongly assumed that there were no two endpoints that
have the same length and the same prefix which is not true.

Added tests that check the error is returned only for fully matching
endpoints with variable mismatches and not for unknown endpoints.