Make mxs::config::Param un-movable
This makes sure that the type cannot be copied or moved once it has been
initialized. This prevents mistakes that would otherwise happen if one
were to attempt to construct a list of Params and then pass them
around. With the current implementation of Param, this must be prevented
as it results in the parameter being removed multiple times from the
Specification that contains it.