MXS-4707: Use ParamRegex for match parameters
This fixes the problem where escaped regular expressions for the
namedserverfilter were not processed correctly and ended up including
the escape characters in the pattern itself. Even with the fix, the
patterns are still compiled twice for every configuration but doing it
this way results in a minimal patch that fixes the problem.

The legacy `match` parameter still uses a ParamString which will be
fixed when it gets removed.