MXS-4742: Include MaxScale log in reports
The report now includes the MaxScale log. The line count is configurable
but defaults to 1000 lines. The reason for a low default value is to
prevent excessive slowdown of the REST-API due to the log
gathering. Having the log in the report helps provide a snapshot of the
current situation of MaxScale.

Also renamed the field names in the generated report. Using only
alphanumerical names makes them easier to use in tools like jq.