MXS-4742: Add server variables to maxctrl report
The server global variables, global status and InnoDB status are now
included in the report generated by MaxCtrl. This will store all the
required information in one place in a consistent format that'll make
analyzing problems faster.

Since querying the server variables can be slow, the retrieval of the
information must be done in two steps. By first collecting the target
servers to query, the MainWorker is blocked only for the duration of the
iteration over the monitor and server lists. The querying itself is done
asynchronously by using the callback mechanism in the HttpResponse which
executes it in the mxs::thread_pool().

The information returned by the endpoint is not something that is
expected to be monitored often, at least for now, which is why it is
prefixed in the /maxscale/debug path.

Added a system test for report generation. In practice the report
generation has already been used by the system tests as it is called
when a test ends up in a timeout. Explicitly testing that it succeeds in
the maxctrl_basic will catch any case where an error will occur.