MXS-4538: Read status back from the cluster
Restoring the status of the local node to the state that's stored in the
cluster fixes the locally stored state that is set to a network error if
a cluster outage occurs.

The value cannot be explicitly set after a restart as it'll store an OK
status in the cluster even if the previous update caused an error.