MXS-4776: Load balance session commands correctly
In the case where a new connection must be opened because a session
command is about to be executed, the server selection always preferred
the first server. This caused a skew in routing to take place if used
with `lazy_connect` and the first command to be executed was a session
command.