Attempt to dump stacktraces on SIGABRT
If the SystemD watchdog kills MaxScale, the most useful thing for
figuring out why it happened is a stacktrace from all threads. This is
already doable with debug=gdb-stacktrace but this requires it to be
enabled before the problem occurs.

With the introduction of the built-in profiler, the same mechanism can
be used to collect stacktraces from all threads in MaxScale. This is a
risk worth taking as watchdog timeouts are very hard to analyze
after-the-fact.