MXS-4750 Match correctly when hostnames are used
If hostnames (and not IPs) are used in the server definitions
and dynamic_node_detection=false is used, then the XpandNodes
will be created directly from the hostnames. Thus, when figuring
out which node corresponse to which server, we need to match the
server address (i.e. hostname) directly to the ip of the node.