MXS-4728 Refactor and simplify `mxs-filter-list`
Selecting all items for dynamic items causes extra computation
steps for the `mxs-virtual-table` component.

This simplifies the filter list component to accept a list of string
items and output "nonticked" items instead of ticked items.

This also adjusts the components that uses this filter list component
to adapt this changes except the `result-data-table` in the query editor
as it will be later refactored to use `mxs-virtual-table`.