MXS-4786: Add function for unexpected situations
The function dumps the current trace log contents if it has been
enabled. If it isn't, a warning is logged instructing the user how to
enable it.

The main use of this function for now is to dump the trace information
whenever unknown PS IDs are encountered. These problems are hard to
debug without the info level logs and having the trace logs is the next
best thing.

This behavior could be extended to automatically turn the trace log on
if the same problem has been encountered often enough. For example, if
the problem occurs twice, the trace log could then be automatically
turned on for a short while. Once the desired amount of trace log dumps
have been done, the trace logging could then be automatically disabled.