MXS-4789 Fix workspace components caching - part 1
Caching in the workspace page doesn't work as expected. Computed
properties data based on the active worksheet or query tab ID changes
even when the component is kept inactive.

To resolve this issue, the ID must be passed via props and computed
properties should get data based on this ID instead of getting data
from the model getters. By following this approach, the usage of
activated and deactivated hooks for manually adding the watchers
is no longer needed.