MXS-4728 Add `filterDateFormat` property for table header
Function can't be pass to web worker without using external
library. i.e. `jsonfn`. So `customGroup` or `customFilter`
functions from the old virtual scroll table can't be included
in the new virtual table. Anyway, those custom functions are
used only in the `HistoryAndSnippetsCtr` component to format
timestamp value before filtering or grouping. There is a simpler
approach which is to provide a date format for the date column.

If there are needs to use more advanced filtering or grouping, `jsonfn`
can be installed and used to overcoming the no-functions limitation
of web worker.