MXS-4787: Optimize commonly logged messages
The use of sprintf in commonly logged messages shows up as taking a
significant amount of CPU time when session_trace is enabled. The other
major reason was std::ostringstream which is very slow compared to
simple std::string concatenation (e.g with mxb::cat).

The two describe() functions show up in profiling as taking more time
than the actual logging function call. Thus optimizing the message
generation there should result in a smaller overhead for in-memory
tracing with session_trace.

In Reply::describe() there remains a large overhead in the call to
mxb::pretty_size() that happens due to the conversion of the size to a
floating point number. Profiling shows that the function call takes 80%
of the measured CPU time after the changes done in this commit.

The mxs_get_context used sprintf for converting the integers into
strings. This is not optimal and benchmarks show that using
std::to_chars is about 6 times faster than using sprintf. With Ubuntu
18.04 no longer being supported, the remaining supported operating
systems all have GCC 8.1 or newer. This means that the <charconv> header
exists and is implemented for integer types.