MXS-4787: Store trace log in mxb::Window
A ring buffer is a good container for storing a moving window of
messages. It removes the need to allocate memory for each new message
and, when compared to a std::deque, slightly improves the locality of
the logged messages by removing one level of indirection.

The mxb::Window is an implementation of such a sliding fixed-size window
that is essentially an append-only ring buffer. The notable change to
the std::deque version is that if the buffer size decreases, the number
of held messages will correctly shrink to the newly requested size. The
old version never shrunk to a smaller size.

Added unit tests for the window size that covers both overflow, partial
windows and resizing via move construction.