Download pcre2 when building
The pcre2 sources have been included with the MaxScale sources,
but that makes updating the version cumbersome and no other
third-party library is handled that way.

This now changes only the way pcre2 is built and the version
10.23 is the same as that of the source (in 2.5). When this commit
is merged upward, the version will in 6.4 be changed to 10.36, which
is the version of pcre2 in 6.4.

Then the existing pcre2 directory will be removed in another
change and finally there will be a change in 2.5 that bumps
the version to 10.42, which then will be used in all branches.