MXS-4764: Add `newest` and `oldest` to `gtid`
The two configuration values allow the replication to be started from
either the beginning of the available history with `oldest` or the very
latest event with `newest`. The `oldest` values allows a contiguous
history to be stored in Kafka even if the start of the history is not in
the binlogs. With `newest` only events generated after the replication
has started get pushed to Kafka.