MXS-4728 Revert the implementation of `mxs-virtual-table`
In hindsight, even after using web-worker to compute the layout of
the grid, it perform poorly for filtering, sorting, resizing
and other operations as it requires to iterate through the each
row and cell to assign new x,y, width, or height value.
The existing approach is significantly faster and only lack
the support of virtualization on the horizontal.

Therefore, it's better to optimize `mxs-virtual-scroll-tbl`
keep the web-worker to optimize some operations.

To workaround the high memory consumption of rendering a large number
of table columns, there should be a limit and a warning text to the
user.

In brief, virtualizing on both directions is possible but using
`vue-virtual-collection` is not a good idea. Besides, the maximum
number of columns is 1000 for InnoDB, and it's not practically to
show all of the columns.

Anyway, the `mxs-virtual-scroll-tbl` still has a memory consumption
issue and should be resolved to have smooth scrolling.