MXS-4728 Show a warning text if columns length > 50
It's not practical to see more than 50 columns at once, so
there should be a limit. i.e. 50 columns. The first 50 columns
will be rendered, the visibility of the rest can be toggled
using the columns dropdown.

This also adjusts the default `bench` props for the vertical
virtual scroller to 1, so that only one row outside the user
view that are rendered.