MXS-4807 Fix mxb::get_release_string()
By default, the version is first looked up from /etc/os-release,
but if the file does not exist (it should on all systemd systems)
another attempt is made with /etc/lsb-release.

The code now only assumes that the file consists of PARAM=VALUE
lines, where the VALUE may be enclosed with double quotes.

The old function is removed in a separate commit as removing it
in this commit caused the diff to become a complete mess.