Change the default name of ETL worksheet task
To keep the naming consistent among worksheet tasks.
This sets the default name to uppercase. So the default
name of the worksheet would be as follows:
`QUERY EDITOR`, `NEW MIGRATION`, `ERD`, `WORKSHEET`.

This also adjusts the tab content's max-width to 120px
so the `NEW MIGRATION` text can be fully visible.