MXS-4789 Fix caching issue on ETL worksheet
This passes down needed data via props and remove
the need for using `activated` and `deactivated` hook
as the `watcher` for etl data won't be re-evaluated
when switching between etl worksheets.

This also optimizes `handleEtlCall` action by removing
the dispatch of `getEtlCallRes` action. The action to
get etl res will be called automatically by the `queryId`
watcher in `EtlMigrationStage` component.