Fix release mode build failure of kafkacdc
The compiler warned about writing past-the-end of the buffer if the
to_chars, for whatever reason, did not convert the number correctly. By
limiting the range to the expected 20 byte size of a 64-bit integer, a
valid range is always used even in the theoretical case where the
number-to-string conversion would fail.