MXS-4814: Add reset-gtid module command
The new reset-gtid command resets the global GTID state. This has the
same effect as restarting MaxScale or destroying the readwritesplit
service would have.

Extended the test case to check that the functionality works. Also added
a new diagnostic field that shows the GTID state that the router is at
currently. Checking that the GTID state shown in the diagnostic output
changes is the simplest and most effective way to test this as it uses
the same function that the actual SQL traffic uses.

Documented the new module commands in the readwritesplit documentation.