MXS-4818 Using `indexes` instead of header names
Using `header.text` for identifying the column will break
some features like sorting, filtering, grouping and visibility
toggling in an edge case where columns have the same name.
e.g.
```
SELECT Host as User, User FROM mysql.user
```

This happens only in an edge case and there was a refactoring
in develop, so fixing it in older release is not convenient.