Add multi-statement test cases to test_pp_sqlite
This is a somewhat small test of the multi-statement detection but it
covers the cases that are expected.

Added testing of the generic is_multi_stmt() version to make sure it
produces the same result as the SIMD version.

One of the test cases also reproduces the crash in MXS-4817 that happens
if trailing whitespace or a trailing comment is left in the SQL string
after a semicolon.