MXS-4728 Conditionally bind cell events to slot content
This renames `searchHighlighterDisabled` property to
`useCellSlot` so it can be used for conditionally
bind events to the slot. When using the cell slot, the cell
events can be added via `on` property which is similar to
vuetify component.

This also uses a reusable external tooltip for the `action`
column of the history table.