Reorganize RWSplitSession::handle_got_target
This is to simplify a following commit that fixes the problem where a
failure in RWBackend::write() does not trigger the retrying of the
query.

By storing the query in m_current_query before dealing with causal
reads, the prefix SQL addition can be simplified into a simple function
that transforms the given buffer. The result of this is that
m_current_query is only modified in one place while routing queries.

The hint addition to the buffer that causes a failing causal read to be
retried on the master is now done in the same place for both COM_QUERY
and COM_STMT_EXECUTE which should help clarify the logic of the code.