Move `store` calculation inside handle_got_target()
This way the variable can be made const and is assigned only once. The
argument can then be replaced with the RoutingPlan which is needed in
the calculation of the value.

The addition of the RoutingPlan to the function also allows the
`target->select_started()` call to use TARGET_IS_SLAVE() which fixes the
problem where reads routed to masters weren't counted as reads. The call
to `select_started()` is now also done even if the session is locked to
the master. This makes the measurement more consistent regardless of the
configuration.