Hide readwritesplit internal symbols
Use -fvisibility=hidden in readwritesplit and only expose the
MXS_CREATE_MODULE function. This cuts down the number of real exported
symbols to just one.

There are still some zlib symbols that get pulled in from the
connector-c library that appear to remain in the final library. This can
be fixed by using the system zlib in the connector.