MXS-4824: Use add_compile_options()
This is a more explicit approach to defining compiler flags instead of
using the older CMAKE_C[XX]_FLAGS which is passed to both the compiler
and the linker.

The flags that are only for C++ or for Debug/RelWithDebInfo
configuration are passed via generator expressions. This seems to be the
recommended approach to pass options that are either configuration or
compiler specific.