MXS-4825: Throw exception on causal read sync failure
This now catches errors that would have previously been undetected and
would've resulted in the synchronization not being done. However, given
the fact that two subsequent calls to RWBackend::write() in the same
function are very unlikely to fail, this is not a very practical
problem.

This also changes how the buffer storage is done. Instead of copying it
into m_current_query, the query is now only stored there if the query is
successfully routed. This way there are no modifications to the internal
state unless the routing was successful.