MXS-4826: Use real module name in nested parameters
This fixes the problem with nested configurations where the protocol
name must exactly match the parameter name. Since the MariaDB client
protocol has several aliases, the names will not match if any of the
aliases is in use. To make this easier to configure, the nested
parameter handling should resolve the module names before using them.