MXS-4017 Remove `completion_items` state
When `identifier_auto_completion` is true, this state
stores only schema names. So if the user sets one
schema as the active and then expand other schemas,
new completion items won't be generated which is bad.

The state will be later replaced with a computed property
that is generated by flatting the schema tree where the
active schema will be excluded.

This also renames some functions in `schemaNodeHelper` and
cleans up obsolete code related to the generation of
completion items.