MXS-4825: Use exceptions in slave failure handling
The handle_error() function now closes the backend and decrements the
expected response count in only one place. This reduces code duplication
as the operations need to be done regardless of the backend
type. Similarly, transaction replay can be handled in a similar manner
as it is agnostic to the server type.