MXS-4017 Show more details about the completion item
This refactors `genNodeGroupSQL` to have an option to return
identifier name and its parent identifier names.

The users can now see the schema name or table name of
a completion item. e.g. If the users type `id` column name,
the suggestion widget will pop up showing the schema name
and table name of that column.