MXS-4803 Detect initial gtid search error
This detects the problem when the Writer does not correctly
read the most recent binlog file. The file could be a corrupted,
but also causes problems when the file was encrypted (MXS-4803).

The problem came from a merge from an older version,
which (obviously) did not decrypt the latest file to find
the most recent gtid. No gtid was found if the file was
encrypted, but this was not an error but rather still allowed
the Writer to start correctly with otherwise readable files.
But these errors would or could happen:
- A crash leaving an incomplete transaction in the file, would
  allow the writer to append new transactions on top of that trx.
- New Readers (slaves) would not be able to properly connect.