Run kill_query in parallel
The test now runs 50 threads in parallel that send a query that is known
to not complete in any reasonable amount of time. This should improve
the test speed slightly but mainly it'll work to better differentiate
the case when the query is killed by the max_statement_time instead of
the KILL command.