Speed up mxs1507_trx_replay
All the test cases can be run in parallel which makes the test complete
in about 8 seconds when it previously took around 118 seconds.

This is also a verification of the transaction replay mechanism in the
sense that it makes sure that there's no side-effects from running
multiple transaction replays in parallel.