MXS-4838: Remove duplicate m_worker from MXS_SESSION
The worker was already stored in the Callable. In addition, the one in
MXS_SESSION was modified earlier than the one in Callable which appears
wrong. Further investigation in the correctness of this is needed and
the only comment in the code refers to a class that has already been
removed.

This also slightly adjusts the runtime code to avoid TLS lookups where
they are not needed. The assumption that the current thread owns the
session is still kept but it is now only enforced with debug assertions
and the code itself uses the pointer stored in the Callable.