MXS-4838: Increment m_num_responses after routeQuery
With the refactoring done for 23.08, the pre-routeQuery increment is no
longer needed as no direct calls to clientReply are made. This makes it
so that the number of expected responses remains zero for the duration
of the routeQuery call after which it will be incremented, if
successful. This also means that the client protocol's is_idle()
function can be used inside routeQuery and it will report the idleness
correctly if no responses are currently expected.

This does require that any responses that will be delivered must be sent
with a lcall() so that the client protocol can return from the
routeQuery call.