MXS-4838: Remove session references from the REST-API
The references can be removed by combining the lookup operation to the
actual work that is to be done if the lookup is successful.

This fixes two long-standing problems: a race condition between the
validity check of the ID and the actual use of it and the overall use of
session references in the REST-API. This is also more efficient as it
completely avoids the two extra calls to the RoutingWorkers that were
previously made to acquire the session references.