MXS-4838: Prevent session restarts in some cases
If `idle_session_pool_time` has been enabled, the restarting of session
is disabled for that service as it installs a dcall that would get
removed by it. Similarly, if delayed routing, a KILL command or any
other command that causes a dcall to be added is in progress the
restarting is not done. This simplifies the restarting at the cost of
not being able to do it in all cases.

The code can currently report a successful restart even if the session
will eventually fail to do it. There's nothing that prevents new dcalls
from being added after the REST-API call has finished and before the
call to do_restart() is done.

Since the dcalls have the timing information in them, it could be used
to determine when the restart can be attempted again. This could even be
implemented as a dcall itself that then does a lcall that in turns calls
Session::restart() to trigger the restart again.