MXS-4838: Invoke lcalls manually in test_cacheoptions
The test relies on the replies to cached queries being delivered
immediately. Now that the queued responses are delivered using a lcall,
the test requires a way to force the lcalls to be delivered without
leaving the function. The simplest way to have the same behavior as
previously is to move the lcall deliver to a function and expose it to
unit tests.