MXS-4838: Store Endpoints in std::shared_ptr
This fixes a potential race condition that can happen when a session is
restarted: when the routing tree is built again, the old Endpoint
objects get deleted if the operation is successful. If there are any
pending dcalls or lcalls while this is happening, it is possible that
the Endpoint pointer they used now points to freed memory.

The solution to this is to use a weak reference to the Endpoint which
requires that they are stored in a std::shared_ptr instead of a
std::unique_ptr.

The Endpoint for the ServiceUpstream was also not correctly set. This
would cause a null pointer dereference for any responses delivered by
the first filter in the chain.