MXS-4839 Add RoutingWorker::[suspend|resume]_sessions()
The suspend_sessions() function suspends those sessions that
can be suspended. I.e. unless the return value indicates that
all sessions have been suspended, it need to be called again
(and again) until there are non non-suspended left.
The resume_session() just resumes those sessions that currently
are suspended.

This is not necessarily the final approach. It might make sense
to make the sessions explicitly aware of the suspending, so that
they could suspend themselves once the condition that prevent
the suspending is no longer present.