Use separate tests in mxs3663_universal_causal_reads
The reuse of the test execution code made it hard to determine the
correctness of the test itself. By separating them into their own test
cases, the expected behavior is visible from the code itself.

Also adjusted the timeouts in the test to fail faster. This should speed
by the test by a tiny bit.