Make MonitorManager::wait_one_tick() more stable
There is a common pattern in the system tests: block a node and then
call the /maxscale/debug/monitor_wait endpoint. This makes it so that
there's no need to wait in the tests for the blocked server to be
noticed by the monitor. This greatly speeds up testing while also making
them more deterministic.

Waiting for one tick won't guarantee that a full monitoring loop has
been done. For most cases, the monitoring interval is long enough that
it doesn't have an effect. However, there is still a possibility that
the monitor only partially executes the loop and any changes to the
environment done by the test code are thus not visible to the rest of
MaxScale.

Waiting that at least two intervals have passed will make sure that the
whole monitoring loop has been done to completion.