Run test in parallel in mxs321
The test checks that exceeding max_connections doesn't cause
problems. There's no need to run each test separately as the number of
created connections always exceeds the limit. By creating a connection
to each service once per loop, the test will have the same effect that
the old one had except that it'll be three times faster.

With these changes, the test takes a few seconds to run instead of the
30 seconds that it took previously.