MXS-4851 Fix services dashboard visual glitch in routing targets column
If there are 2 routing targets displayed in the `Services` dashboard
table, the comma separator isn't added making it difficult to read.

The bug exists since the introduction of the `v-mxs-highlighter` directive
which overwritten the text content of the element using the directive.

This adjusts the condition so that if there are more than 1 routing
targets, it will be shown in an interactive tooltip where each line in
the tooltip is a routing target.