MXS-4572 Add `widthRelationship` props to `server-form-input` component
Form components in `CreateResource` will be later reused for
creating MaxScale Config Wizard. All form inputs can be reused
immediately except the `server-form-input`. The wizard will
guide the users to firstly create the servers, monitors, filters,
services and finally the listeners. Because the servers will be
created in the first step of the wizard thus there is no need to
specify the monitor in the first step.

This adds the mentioned props to conditionally show the relationship
selection inputs and output form values.