MXS-4572 Split `defaultItems` into separate props
The `ServiceFormInput` component has two relationship types,
therefore, it's better to have dedicated props for certain types.
Later on, in the wizard UI, when the users enter to `Service` stage,
the recently created servers and filters will be the default relationship
items.

This also cleans up and simplifies some methods in `CreateMxsObj` and
removes obsolete getters.