MXS-4856 Prevent monitored server being selected as the default item
When the users are on the server details page, if they create
a new monitor, the current server name will be selected as the
default server in the monitor's relationship. The `defaultItems`
props doesn't filter out monitored server. As a result, the API
returns the following error:
```
"Server 'server_name' is already monitored by 'monitor_name',
cannot add it to another monitor.Could not create monitor
'new_monitor' with module 'mariadbmon'"
```

This adds a computed property called `defServers` for filtering
`defaultItems` to keep only non-monitored servers.