MXS-3986 Change the poll() call to always wait max x seconds
The compression thread updates files, which causes the main
thread to wake up possibly before the compression thread is
done. This can lead to the main thread not seeing compression
as done, and pick up the next file to compress.
Since there already is a timeout to perform the next purge,
that timeout is now set to a fixed 2 seconds to guarantee
the next file to compress is picked up even if there is no
other activity.

Also move the check whether a purge should be done into the purge function.
Avoids calling Clock::now() twice.