MXS-4572 Add `showAdvanceToggle` props to `ModuleParameters`
If the props value is true, a toggle input will be shown to toggle
`isAdvanced` state. The default value os `isAdvanced` is false, so
only mandatory parameters will be shown with an exception for
`address, port, socket` parameters of server and listener type.

This also adds `hideModuleOpts` props and take it
into use for `ServerFormInput`.