MXS-4761 Expose get_object_type()
If modules create new objects they need to be able to check
whether some object with that name exists already.

From a belt and suspenders perspective it should simply not be
possible to create an object using a name that already is used,
but at least as far as services go that is currently possible.