MXS-4798 Avoid emergency promotion if master has a broke slave
Previously, the monitor would go into new master selection mode
if current master goes down and it had no properly replicating
slaves. In this mode, a server with broken replication could be
selected as new master.

This commit changes the logic such that the monitor sticks with
the old master as long as it has even a partially broken slave
(either IO or SQL thread is active). Also, a server with a
broken replication connection to another monitored server cannot
be selected as master unless server is part of a cycle.