MXS-4572 Passing down the type of the object via props
The `ServerFormInput` has been recently refactored to use
the `ModuleParameters` instead of the `ParametersCollapse`
component. As a result, it shows the `type` parameter which
is not needed when creating a new server object. That parameter
is received from the API and shouldn't be shown to the user.

This also removes the `isListener` props as it can be replaced
with the `objType` props.